Title: 평균점수 구하기, Time: 158ms, Memory: 12MB, Status: Passed - Codetree

+++ 250614/평균점수 구하기/print-average.java
@@ -7,7 +7,7 @@ public class Main {
     for (int i = 0; i < score; i++){
         total += sc.nextDouble();
     }
-        System.out.printf("%.1f\n",(float)total / score);
+        System.out.printf("%.1f\n",total / score);
 
     }
 }
